+++ test/lazy_stream_test.dart
@@ -83,7 +83,7 @@ void main() {
   });
 
   test("a lazy stream can't be listened to multiple times", () {
-    var stream = LazyStream(expectAsync0(() => Stream.empty()));
+    var stream = LazyStream(expectAsync0(Stream.empty));
     expect(stream.isBroadcast, isFalse);
 
     stream.listen(null);